+++ ExampleMod/Content/Items/Placeable/Banners/ExampleCustomAISlimeNPCBanner.cs
@@ -1,0 +1,18 @@
+﻿using ExampleMod.Content.Tiles.Banners;
+using Terraria.Enums;
+using Terraria.ModLoader;
+using Terraria;
+
+namespace ExampleMod.Content.Items.Placeable.Banners
+{
+	public class ExampleCustomAISlimeNPCBanner : ModItem
+	{
+		public override void SetDefaults() {
+			Item.DefaultToPlaceableTile(ModContent.TileType<EnemyBanner>(), (int)EnemyBanner.StyleID.ExampleCustomAISlimeNPC);
+			Item.width = 10;
+			Item.height = 24;
+			Item.SetShopValues(ItemRarityColor.Blue1, Item.buyPrice(silver: 10));
+		}
+	}
+	
+}
